+++ src/config.rs
@@ -79,13 +79,13 @@ pub fn write_default_file_if_missing() {
 
     let config = Config {
         default: ModeConfig {
-            inactive: (BooleanLightMode::Off, RedAmberGreenLightMode::Green),
+            inactive: (BooleanLightMode::On, RedAmberGreenLightMode::Green),
             active: (BooleanLightMode::On, RedAmberGreenLightMode::Amber),
             blocked: (BooleanLightMode::Off, RedAmberGreenLightMode::Red),
             alert: (BooleanLightMode::Flash, RedAmberGreenLightMode::RedAmber),
         },
         hardpoints_deployed: ModeConfig {
-            inactive: (BooleanLightMode::Off, RedAmberGreenLightMode::Red),
+            inactive: (BooleanLightMode::On, RedAmberGreenLightMode::Red),
             active: (BooleanLightMode::On, RedAmberGreenLightMode::Amber),
             blocked: (BooleanLightMode::Off, RedAmberGreenLightMode::Off),
             alert: (BooleanLightMode::Flash, RedAmberGreenLightMode::RedAmber),